Don't use callbacks for the test framework
Using callbacks makes our functions "impure". We don't know what happens
inside of them, and we don't control it. By moving the "setting up" of
the test to the caller, we ensure that, if test() is called, it will work.
Errors will happen before it gets added to the callstack, moving it out of
wunit's are of responsibility.

The tradeoff is a slightly odd-looking syntax using self-invoking anonymous
functions, but it remains readable and understandable.

Using the same logic for define() also helps, although define() remains impure,
as it prints to the console. Still, if define() is picked from the callstack,
the chance of it working is pretty high, as we only depend on the console
working, not any passed callbacks.